Add end-to-end testing
