+++ compiler/dafny/tests/.gitignore
@@ -1,0 +1,3 @@
+# Generated by Dafny
+*.dll
+*.runtimeconfig.json
