+++ compiler/dafny/tests/hello.dfy
@@ -1,0 +1,4 @@
+method Main()
+{
+  print "Hello, Cake";
+}
